Getting the latest updates of the project

+++ project/DataSource/UnitOfWork.js
@@ -10,6 +10,7 @@ let DimensionsTDG = require(rootPath + '/DataSource/TableDataGateway/DimensionsT
 let LaptopsTDG = require(rootPath + '/DataSource/TableDataGateway/LaptopsTDG.js');
 let MonitorsTDG = require(rootPath + '/DataSource/TableDataGateway/MonitorsTDG.js');
 let TabletsTDG = require(rootPath + '/DataSource/TableDataGateway/TabletsTDG.js');
+
 class UnitOfWork {
   constructor() {
     this.environment = process.env.NODE_ENV || 'development';
@@ -51,8 +52,10 @@ class UnitOfWork {
       console.log(this.dirtyElements[0]);
       console.log("Electronics to delete: ");
       console.log(this.deletedElements[0]);
-
-      Promise.each(this.deletedElements[0], (electronics) => {
+      let deletedItems;
+      //delete items
+      if(this.deletedElements[0].length > 0){
+       deletedItems = Promise.each(this.deletedElements[0], (electronics) => {
         return this.InventoryItemsTDG.delete(electronic, trx).then((serial_number) => {
           Promise.each(electronic.serial_number, (item_serial_number) => {
             return this.deleteInventoryItems(item_serial_number, electronic.serial_number).then((id) => {
@@ -60,10 +63,56 @@ class UnitOfWork {
             })
           })
         })
-      })
+      });}//.then(trx.commit).catch(trx.rollback);
+      //end of delete
 
-      Promise.each(this.newElements[0], (electronic) => {
-        //console.log("Serial number"+ electronic.serial_number);
+      //update
+      let updateditems = Promise.each(this.dirtyElements[0], (electronic) => {
+         this.productDescTDG.update(electronic, trx).then((model_number) => {
+          /*Promise.each(electronic.model_number, (item_model_number) => {
+            return this.updateDescription(item_model_number, electronic.model_number).then((id) => {
+              console.log('updated item description');
+            });
+          });*/
+
+          switch (electronic.type) {
+            case 'Desktop':
+              {
+                return this.dimensionsTDG.update(electronic).transacting(trx).then((dimensionsId) => {
+                  return this.computersTDG.update(electronic).transacting(trx).then((compId) => {
+                    return this.desktopsTDG.update(compId, dimensionsId, electronic).transacting(trx);
+                  });
+                });
+              };
+              break;
+            case 'Laptop':
+              {
+                return this.computersTDG.update(electronic).transacting(trx).then((compId) => {
+                  return this.laptopsTDG.update(compId, electronic).transacting(trx);
+                });
+              };
+              break;
+            case 'Tablet':
+              {
+                return this.dimensionsTDG.update(electronic).transacting(trx).then((dimensionsId) => {
+                  return this.computersTDG.update(electronic).transacting(trx).then((compId) => {
+                    return this.tabletsTDG.update(compId, dimensionsId, electronic).transacting(trx);
+                  });
+                });
+              };
+              break;
+            case 'Monitor':
+              {
+                return this.monitorsTDG.update(electronic).transacting(trx);
+              };
+              break;
+          }
+        });
+      });//.then(trx.commit).catch(trx.rollback);
+      //end of update 
+
+      //add items
+      let addeditems = Promise.each(this.newElements[0], (electronic) => {
         return this.productDescTDG.add(electronic, trx).then((model_number) => {
           Promise.each(electronic.serial_number, (item_serial_number) => {
             return this.addInventoryItem(item_serial_number, electronic.model_number).then((id) => {
@@ -106,7 +155,8 @@ class UnitOfWork {
         });
       }
       //add
-      ).then(trx.commit).catch(trx.rollback);
+      );//.then(trx.commit).catch(trx.rollback);
+      Promise.all([deletedItems, updateditems, addeditems]).then(trx.commit).catch(trx.rollback);
     });
   }
 
@@ -114,6 +164,10 @@ class UnitOfWork {
     return this.connection.from('Inventory').where('id', id).del({'serial_number': serial_number});
   }
 
+
+  updateDescription(){
+
+  }
   addInventoryItem(serial_number, model_number) {
     return this.connection.insert({
       'model_number': model_number,
